% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say}
\alias{say}
\title{Summmon a whale for a customised pick-me-up}
\usage{
say(what)
}
\arguments{
\item{what}{The phrase the whale will utter}
}
\value{
Whale shaped message
}
\description{
Summmon a whale for a customised pick-me-up
}
\examples{
say("You're whale-come")
}
