% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohwhaley2.R
\name{say}
\alias{say}
\title{ohwhaley2}
\usage{
say(what)
}
\arguments{
\item{what}{What you want the whale to say}
}
\value{
Text string with pun
}
\description{
ohwhaley2
}
\examples{
say("You're whale-come")
}
